refactor: move calculator in new class
