+++ Calculator.monsieur-raptor/Calculator.monsieur-raptor.slnx
@@ -1,3 +1,4 @@
 <Solution>
   <Project Path="Calculator.monsieur-raptor/Calculator.monsieur-raptor.csproj" />
+  <Project Path="CalculatorLibrary/CalculatorLibrary.csproj" />
 </Solution>
